requirements.txt
antlr4-python3-runtime
oletools
pandas
tqdm

llama-cpp-python
ollama
transformers